code fromat and run cargo fix

+++ spi/spi-conf/src/serv.rs
@@ -241,7 +241,7 @@ fn jwt_token_key(token: &str) -> String {
 async fn bind_token_ctx(token: &str, ttl: u64, ctx: &TardisContext) -> TardisResult<()> {
     let funs = crate::get_tardis_inst_ref();
     let ctx_json = tardis::basic::json::TardisJson.obj_to_string(ctx)?;
-    funs.cache().set_ex(&jwt_token_key(token), &ctx_json, ttl as usize).await.map_err(|e| TardisError::internal_error(&format!("{e}"), CACHE_ERROR))
+    funs.cache().set_ex(&jwt_token_key(token), &ctx_json, ttl as u64).await.map_err(|e| TardisError::internal_error(&format!("{e}"), CACHE_ERROR))
 }
 
 /// get the tardis context by jwt token
